requirements.txt
pyQT5
numpy
opencv-python
Ipython
keyboard>=0.13.5
pyusb>=1.2.1
pyserial>=3.5